# Conflicts:
#	.idea/compiler.xml
#	server/.gradle/7.6.1/checksums/checksums.lock
#	server/.gradle/7.6.1/dependencies-accessors/dependencies-accessors.lock
#	server/.gradle/7.6.1/executionHistory/executionHistory.bin
#	server/.gradle/7.6.1/executionHistory/executionHistory.lock
#	server/.gradle/7.6.1/fileHashes/fileHashes.bin
#	server/.gradle/7.6.1/fileHashes/fileHashes.lock
#	server/.gradle/7.6.1/fileHashes/resourceHashesCache.bin
#	server/.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	server/.gradle/buildOutputCleanup/cache.properties
#	server/.gradle/file-system.probe
#	server/build/classes/java/main/pro/stackOverFlow/answer/controller/AnswerController.class
#	server/build/classes/java/main/pro/stackOverFlow/exception/ExceptionCode.class
#	server/build/classes/java/main/pro/stackOverFlow/member/entity/Member.class
#	server/build/generated/sources/annotationProcessor/java/main/pro/stackOverFlow/member/mapper/MemberMapperImpl.java
#	server/build/tmp/compileJava/previous-compilation-data.bin
#	server/src/main/java/pro/stackOverFlow/answer/controller/AnswerController.java
#	server/src/main/java/pro/stackOverFlow/answer/entity/Answer.java
#	server/src/main/java/pro/stackOverFlow/answer/service/AnswerService.java
#	server/src/main/java/pro/stackOverFlow/exception/ExceptionCode.java
#	server/src/main/java/pro/stackOverFlow/member/entity/Member.java
#	server/src/main/java/pro/stackOverFlow/question/controller/QuestionController.java
#	server/src/main/java/pro/stackOverFlow/question/entity/Question.java
#	server/src/main/java/pro/stackOverFlow/question/service/QuestionService.java

+++ server/src/main/java/pro/stackOverFlow/question/service/QuestionService.java
@@ -1,26 +1,28 @@
 package pro.stackOverFlow.question.service;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import pro.stackOverFlow.member.entity.Member;
 import pro.stackOverFlow.question.entity.Question;
 import pro.stackOverFlow.exception.BusinessLogicException;
 import pro.stackOverFlow.exception.ExceptionCode;
+import pro.stackOverFlow.question.entity.QuestionVote;
 import pro.stackOverFlow.question.repository.QuestionRepository;
+import pro.stackOverFlow.question.repository.QuestionVoteRepository;
 
 import java.util.Optional;
 
 @Service
 @Transactional
+@RequiredArgsConstructor
 public class QuestionService {
 
     private final QuestionRepository questionRepository;
-
-    public QuestionService(QuestionRepository questionRepository) {
-        this.questionRepository = questionRepository;
-    }
+    private final QuestionVoteRepository questionVoteRepository;
 
     // create
     public Question createQuestion(Question question) {
@@ -30,8 +32,6 @@ public class QuestionService {
     // read
     public Question findQuestion(Long questionId) {
         Question question = findVerifiedQuestion(questionId);
-        question.setViewCount(question.getViewCount() + 1);
-
         return questionRepository.save(question);
     }
 
@@ -47,16 +47,65 @@ public class QuestionService {
     }
 
     // delete
-    public void deleteQuestion(Long questionId) {
+    public void deleteQuestion(Long memberId, Long questionId) {
         Question question = findVerifiedQuestion(questionId);
+        Long findMemberId = question.getMember().getMemberId();
+        if (!memberId.equals(findMemberId)) {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_EXISTS);
+        }
         questionRepository.delete(question);
     }
 
+    // question 있는지 검증
     private Question findVerifiedQuestion(Long questionId) {
         Optional<Question> optionalQuestion = questionRepository.findById(questionId);
         return optionalQuestion.orElseThrow(() -> new BusinessLogicException(ExceptionCode.QUESTION_NOT_FOUND));
     }
 
+    // 조회수 카운트
+    public void addViewCount(Question question) {
+        question.setViewCount(question.getViewCount() + 1);
+    }
 
+    // vote
+    public void upVote(Member member, Question question) {
+        QuestionVote questionVote = findQuestionVote(member, question);
+
+        if (questionVote.getVoteStatus().toString().equals("UP")) {
+            throw new BusinessLogicException(ExceptionCode.ALREADY_UP_VOTED);
+        } else if (questionVote.getVoteStatus().toString().equals("NONE")) {
+            questionVote.setVoteStatus(QuestionVote.VoteStatus.UP);
+        } else if (questionVote.getVoteStatus().toString().equals("DOWN")) {
+            questionVote.setVoteStatus(QuestionVote.VoteStatus.NONE);
+        }
+        question.setQuestionVoteCount(question.getQuestionVoteCount() + 1);
+    }
+
+    public void downVote(Member member, Question question) {
+        QuestionVote questionVote = findQuestionVote(member, question);
+
+        if (questionVote.getVoteStatus().toString().equals("DOWN")) {
+            throw new BusinessLogicException(ExceptionCode.ALREADY_DOWN_VOTED);
+        } else if (questionVote.getVoteStatus().toString().equals("NONE")) {
+            questionVote.setVoteStatus(QuestionVote.VoteStatus.DOWN);
+        } else if (questionVote.getVoteStatus().toString().equals("UP")) {
+            questionVote.setVoteStatus(QuestionVote.VoteStatus.NONE);
+        }
+        question.setQuestionVoteCount(question.getQuestionVoteCount() - 1);
+    }
+
+    private QuestionVote findQuestionVote(Member member, Question question) {
+        Optional<QuestionVote> questionVote = questionVoteRepository.findByMemberAndQuestion(member, question);
+        return questionVote.orElseGet(() -> createVote(member, question));
+    }
+
+    private QuestionVote createVote(Member member, Question question) {
+        QuestionVote questionVote = QuestionVote.builder()
+                .voteStatus(QuestionVote.VoteStatus.NONE)
+                .member(member)
+                .question(question)
+                .build();
+        return questionVoteRepository.save(questionVote);
+    }
 
 }
